+++ pkg/commands/apisecurity/operations/addtags.go
@@ -9,8 +9,8 @@ import (
 	"os"
 	"path/filepath"
 
-	"github.com/fastly/go-fastly/v13/fastly"
-	"github.com/fastly/go-fastly/v13/fastly/apisecurity/operations"
+	"github.com/fastly/go-fastly/v14/fastly"
+	"github.com/fastly/go-fastly/v14/fastly/apisecurity/operations"
 
 	"github.com/fastly/kingpin"