+++ pkg/commands/auth/sso_test.go
@@ -7,7 +7,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/fastly/go-fastly/v13/fastly"
+	"github.com/fastly/go-fastly/v14/fastly"
 
 	"github.com/fastly/cli/pkg/auth"
 	authcmd "github.com/fastly/cli/pkg/commands/auth"