+++ pkg/commands/ngwaf/rule/update.go
@@ -9,9 +9,9 @@ import (
 	"os"
 	"path/filepath"
 
-	"github.com/fastly/go-fastly/v13/fastly"
-	"github.com/fastly/go-fastly/v13/fastly/ngwaf/v1/rules"
-	"github.com/fastly/go-fastly/v13/fastly/ngwaf/v1/scope"
+	"github.com/fastly/go-fastly/v14/fastly"
+	"github.com/fastly/go-fastly/v14/fastly/ngwaf/v1/rules"
+	"github.com/fastly/go-fastly/v14/fastly/ngwaf/v1/scope"
 
 	"github.com/fastly/cli/pkg/argparser"
 	fsterr "github.com/fastly/cli/pkg/errors"